+++ CHANGELOG.md
@@ -10,12 +10,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Add `.devcontainer` configuration for Visual Studio Code.
 - Add Scrapy version 2.12.0 in the requirements file.
 - Add `logs` and `outputs` directory.
-- Add scrapy framework
+- Add scrapy framework.
+- Add python-dotenv version 1.0.1 in the requirements file.
+- Add `YugiohDBCardsSpider` spider to scrap card details from `yugioh-card.com`.
 
 ### Changed
 - Update project description in the readme file.
 - Add support section in the readme file.
 - Update .gitignore to ignore `logs` and `outputs` directory.
+- Update `settings.py` to read key-value pairs from a `.env` file.
 
 ## [0.0.1] - 2024-11-26